Fix detection of movies whose VWCF/EMPO resource numbers are not defaults
The Director engine itself uses an indexed lookup for these
resources instead of using a fixed value. For the purposes of
detection it is enough to know that they simply exist; it will
probably be necessary to actually modify how MacResourceFile works
to enable the indexed-based lookups at some point, but for now it
is possible to get away with scanning the map for anything with a
given OSType for the purpose of hacking together something that
works enough to start implementing resource decoding.