Switch file selection to use QString most of the time
I had hoped to convert this entirely to use QString and OsString,
given the compatibility issues of holding paths in a plain String,
but since QString holds UTF-16 internally there is no easy way to
convert directly to OsString. This can be revised later as it
should mostly be an edge case.